[api] 백엔드에서 떨어지는 data 의 type이 제네릭으로 변경되며 data.item 안에 묶여서 떨어지는 부분을 수정
