+++ web/src/domains/community/store/index.js
@@ -1,9 +1,9 @@
 /**
- * events index;
+ * community index;
  */
 
 export * from "./slices.js";
-export * from "./thunks.js";
+export * from "./thunks/index.js";
 export * from "./endpoints.js";
 export * from "./dto.js";
 export * from "./api.js";
