+++ web/src/domains/community/store/thunks/index.js
@@ -1,0 +1,8 @@
+/**
+ * community/thunks index;
+ */
+
+export * from "./boardThunks.js";
+export * from "./postThunks.js";
+export * from "./tagThunks.js";
+export * from "./userThunks.js";
